requirements.txt
fastapi
uvicorn
sqlalchemy
celery
httpx
pydantic
pydantic-settings
python-dotenv
redis
asgiref
asyncpg